Move more taks to core_tasks.
Changed a few names, and put each task in its own file,
for uniformity.

Also replaced the first=True argument to run_prep_command's
install with a comment explaining that it must be first in its
register.py, and another comment in extension_loader explaining
that core_tasks must be registered first.

The reason is that first=True was no less fragile - if anything
else registered into the 'test' goal with first=True, it would
preempt run_prep_command, with no comment. At least this way
the comments are clear, and we can maybe (??) move towards
deprecating first=True, or at least relying on it less.
