+++ src/python/pants/backend/core/register.py
@@ -6,7 +6,6 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
                         unicode_literals, with_statement)
 
 import os
-import sys
 
 from pants.backend.core.from_target import FromTarget
 from pants.backend.core.targets.dependencies import Dependencies
@@ -15,7 +14,6 @@ from pants.backend.core.targets.prep_command import PrepCommand
 from pants.backend.core.targets.resources import Resources
 from pants.backend.core.tasks.bash_completion import BashCompletionTask
 from pants.backend.core.tasks.builddictionary import BuildBuildDictionary
-from pants.backend.core.tasks.clean import Cleaner, Invalidator
 from pants.backend.core.tasks.cloc import CountLinesOfCode
 from pants.backend.core.tasks.confluence_publish import ConfluencePublish
 from pants.backend.core.tasks.deferred_sources_mapper import DeferredSourcesMapper
@@ -30,9 +28,6 @@ from pants.backend.core.tasks.markdown_to_html import MarkdownToHtml
 from pants.backend.core.tasks.minimal_cover import MinimalCover
 from pants.backend.core.tasks.pathdeps import PathDeps
 from pants.backend.core.tasks.paths import Path, Paths
-from pants.backend.core.tasks.reporting_server import KillServer, RunServer
-from pants.backend.core.tasks.roots import ListRoots
-from pants.backend.core.tasks.run_prep_command import RunPrepCommand
 from pants.backend.core.tasks.sorttargets import SortTargets
 from pants.backend.core.tasks.targets_help import TargetsHelp
 from pants.backend.core.wrapped_globs import Globs, RGlobs, ZGlobs
@@ -92,31 +87,6 @@ def register_goals():
 
   task(name='builddict', action=BuildBuildDictionary).install()
 
-  # Cleaning.
-  invalidate = task(name='invalidate', action=Invalidator)
-  invalidate.install().with_description('Invalidate all targets.')
-
-  clean_all = task(name='clean-all', action=Cleaner).install()
-  clean_all.with_description('Clean all build output.')
-  clean_all.install(invalidate, first=True)
-
-  class AsyncCleaner(Cleaner):
-    def execute(self):
-      print('The `clean-all-async` goal is deprecated and currently just forwards to `clean-all`.',
-            file=sys.stderr)
-      print('Please update your usages to `clean-all`.', file=sys.stderr)
-      super(AsyncCleaner, self).execute()
-  clean_all_async = task(name='clean-all-async', action=AsyncCleaner).install().with_description(
-      '[deprecated] Clean all build output in a background process.')
-  clean_all_async.install(invalidate, first=True)
-
-  # Reporting.
-  task(name='server', action=RunServer, serialize=False).install().with_description(
-      'Run the pants reporting server.')
-
-  task(name='killserver', action=KillServer, serialize=False).install().with_description(
-      'Kill the reporting server.')
-
   task(name='markdown', action=MarkdownToHtml).install('markdown').with_description(
       'Generate html from markdown docs.')
 
@@ -147,13 +117,6 @@ def register_goals():
 
   task(name='sort', action=SortTargets).install().with_description(
       'Topologically sort the targets.')
-
-  task(name='run_prep_command', action=RunPrepCommand).install('test', first=True).with_description(
-      "Run a command before tests")
-
-  # Source tree information.
-  task(name='roots', action=ListRoots).install('roots').with_description(
-    "Print the workspace's source roots and associated target types.")
 
   task(name='cloc', action=CountLinesOfCode).install('cloc').with_description(
     "Print counts of lines of code.")
