+++ src/python/pants/core_tasks/BUILD
@@ -5,6 +5,11 @@ python_library(
   name='core_tasks',
   sources=globs('*.py'),
   dependencies=[
+    'src/python/pants/base:exceptions',
+    'src/python/pants/base:workunit',
+    'src/python/pants/binaries:binary_util',
     'src/python/pants/goal:task_registrar',
+    'src/python/pants/reporting',
     'src/python/pants/task',
+    'src/python/pants/util:dirutil',
   ])
