+++ tests/python/pants_test/core_tasks/test_roots.py
